api/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
imutils==0.5.3
numpy==1.19.1
opencv-contrib-python==4.2.0.34
Pillow==7.2.0
pkg-resources==0.0.0
pytesseract==0.3.5
python-anticaptcha==0.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
